Separated MenuItem to its own file Move Menu into `core_modules` Meanio adds menu to itself
