+++ lib/core_modules/menu/MenuItem.js
@@ -1,19 +1,6 @@
 'use strict';
 var _ = require('lodash');
 
-//helper functions
-function extractNames (v) {
-  return v.name;
-}
-
-function get_get (roles,v) {
-  return v.get(roles);
-}
-
-function remove_nulls (v) {
-  return v;
-}
-
 //MenuItem class
 function MenuItem (options) {
   options = _.assign ({name: null, title:null, link:null, roles:null}, options);
@@ -27,13 +14,9 @@ function MenuItem (options) {
   this.submenus = options.submenus || [];
 }
 
-function mapDoStrip (v) {
-  return v ? v.strip() : undefined;
-}
-
 MenuItem.prototype.strip = function () {
   return {
-    name: this.name, 
+    name: this.name,
     title:this.title,
     link: this.link,
     roles:this.roles,
@@ -118,57 +101,21 @@ MenuItem.prototype.add = function (mi) {
   return itm;
 };
 
-var allMenus = new MenuItem (),
-  _ = require('lodash');
-function Menus() {
+//helper functions
+function extractNames (v) {
+  return v.name;
 }
 
-function supportMenus(Meanio){
-
-
-  Menus.prototype.add = function(options) {
-    if (arguments.length === 0) return this;
-    if (options instanceof Array) {
-      options.forEach( Menus.prototype.add.bind(this) );
-      return this;
-    }
-    if (arguments.length > 1) {
-      Array.prototype.forEach.call(arguments,this.add.bind(this));
-      return this;
-    }
-    
-    //fixes scaffolding: menu=path
-    if(options.menu !== undefined) {
-      options.path = options.menu;
-    }
-    
-    options = _.assign({
-      path : 'main', 
-      roles: ['anonymous'],
-    },
-    options);
-    options.path = options.path.replace(/^\//, '');
-    var item = allMenus.findOrCreate(options.path.split('/'));
-    item.add(new MenuItem(options));
-    return this;
-  };
-
-  Menus.prototype.get = function(options) {
-    options = options || {};
-    options.menu = options.menu || 'main';
-    options.roles = options.roles || ['anonymous'];
-    options.defaultMenu = options.defaultMenu || [];
-
-    var sm = allMenus.get(options.roles, options.menu.split('/'));
-    if (!sm) {
-      //no menu at all
-      return options.defaultMenu;
-    }
-    var ret = sm.get(options.roles);
-    return ret ? options.defaultMenu.concat(ret.submenus.map(mapDoStrip)) : options.defaultMenu; 
-  };
-  Meanio.prototype.Menus =  Menus;
-
+function get_get (roles,v) {
+  return v.get(roles);
 }
 
-module.exports = supportMenus;
+function remove_nulls (v) {
+  return v;
+}
+
+function mapDoStrip (v) {
+  return v ? v.strip() : undefined;
+}
+
+module.exports = MenuItem;
