+++ lib/mean.js
@@ -44,7 +44,7 @@ Meanio.onInstance = function(func){
   Meanio.instanceWaiters.push(func);
 };
 
-(require('./menu'))(Meanio);
+Meanio.prototype.Menus = require('./core_modules/menu');
 (require('./core_modules/module'))(Meanio);
 (require('./core_modules/db'))(Meanio);
 (require('./core_modules/server'))(Meanio);
